Add compile::features module
This is only moving things around; compile_ctx.rs is huge, and I would
like to add additional helper types related to tracking features in the
near future.
